viperaveil-main/Discord-SNEK/requirements.txt
asyncio
asyncpg
psycopg2
nest_asyncio
python-decouple
py-cord
PyNaCl
requests
bs4
aiohttp
sqlalchemy
sqlalchemy-orm
aiosqlite
beautifulsoup4
spotify
tekore
wavelink
youtube-search-python
git+https://github.com/theGunner295/soundcloud-lib.git
sponsorblock.py
youtube-dl
pydub